+++ src/main/java/com/example/backend/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain;
+package com.example.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
