+++ src/main/java/com/example/backend/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.domain.users.repository;
+package com.example.backend.repository;
 
-import com.example.backend.domain.users.entity.User;
+import com.example.backend.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
